fix: fix the project according to the requirements
