+++ public/index.html
@@ -4,18 +4,9 @@
     <meta charset="utf-8" />
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-
     <title>TS-Cinema</title>
   </head>
   <body>
-    <div
-      id="root"
-      style="
-        min-height: 100vh;
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-      "
-    ></div>
+    <div id="root"></div>
   </body>
 </html>
