+++ src/components/Header/Header.tsx
@@ -1,7 +1,10 @@
 import { FC } from "react";
 import { useSelector, useDispatch } from "react-redux";
+import { RootState } from "../../redux/reducer/rootReducer";
+import { toggleShowModal } from "../../redux/reducer/modalReducer";
 import { Nav, Navbar, Container, Form, Image } from "react-bootstrap";
 import { NavLink } from "react-router-dom";
+import { USER_ICON, ADMIN_ICON } from "../../data/icon";
 import {
   ABOUTUS_ROUTE,
   ADMIN_ROUTE,
@@ -9,14 +12,12 @@ import {
   FILMS_ROUTE,
   LOGIN_ROUTE,
 } from "../../utils/constsRoutes";
-import { RootState } from "../../redux/reducer/rootReducer";
-import { toggleShowModal } from "../../redux/reducer/modalReducer";
 import "./Header.scss";
 
 const Header: FC = () => {
   const { isAuth, role } = useSelector((state: RootState) => state.user);
   const dispatch = useDispatch();
-  const logout = (): void => {
+  const logout = () => {
     dispatch(toggleShowModal(true));
   };
 
@@ -51,17 +52,13 @@ const Header: FC = () => {
 
           {isAuth ? (
             <div className="d-flex justify-content-between align-items-center ">
-              <div onClick={() => logout()} className="link button">
+              <div onClick={logout} className="link button">
                 LogOut
               </div>
               <NavLink to={role === "ADMIN" ? ADMIN_ROUTE : BASKET_ROUTE}>
                 <Image
                   className="user-icon"
-                  src={
-                    role === "ADMIN"
-                      ? "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQG859lMWMgKWIw_iHz-7K5k77ZvxXkDCWLdw&usqp=CAU"
-                      : "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTxk14wu-R6Mv1zP9ObWB9A08P8wra_wxI3JQ&usqp=CAU"
-                  }
+                  src={role === "ADMIN" ? USER_ICON : ADMIN_ICON}
                   rounded
                 />
               </NavLink>
