--- src/components/Registration/index.ts
@@ -1,3 +1,0 @@
-import Registeration from "./Registeration";
-
-export default Registeration;
