+++ src/data/cinemaImage.ts
@@ -1,0 +1,6 @@
+export const IMAGE_CINEMA_MINSK =
+  "https://mag-cinema.com/image/catalog/News/2020/mammut2.jpg";
+export const IMAGE_CINEMA_VITEBSK =
+  "https://image.freepik.com/free-photo/cinema-cinema-attributes-cinemas-films-online-viewing-popcorn-and-glasses_99433-1577.jpg";
+export const IMAGE_CINEMA_BREST =
+  "http://visit-petersburg.ru/media/uploads/tourobject/198842/198842_cover.jpeg.1050x700_q95_crop_upscale.jpg";
