+++ src/data/icon.ts
@@ -1,0 +1,4 @@
+export const ADMIN_ICON =
+  "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTxk14wu-R6Mv1zP9ObWB9A08P8wra_wxI3JQ&usqp=CAU";
+export const USER_ICON =
+  "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQG859lMWMgKWIw_iHz-7K5k77ZvxXkDCWLdw&usqp=CAU";
