+++ src/http/userApi.ts
@@ -1,11 +1,12 @@
 import { $authHost, $host } from "./index";
 import jwt_decode from "jwt-decode";
 import { IUserGet } from "../types/user";
+import { USER } from "../utils/constsRoles";
 
 export async function registration(
   email: string,
   password: string,
-  role = "USER"
+  role = USER
 ): Promise<IUserGet> {
   const { data } = await $host.post("api/user/registration", {
     email,
@@ -28,7 +29,7 @@ export async function login(
   return jwt_decode(data.token);
 }
 
-export async function check(): Promise<IUserGet> {
+export async function checkIsLogin(): Promise<IUserGet> {
   const { data } = await $authHost.get("api/user/auth");
   localStorage.setItem("token", data.token);
   return jwt_decode(data.token);
