+++ src/pages/BuyTicket.tsx
@@ -7,22 +7,20 @@ import Hall from "../components/Hall/Hall";
 
 const BuyTicket: FC = () => {
   const { filmId } = useSelector((state: RootState) => state.filmId);
+  const film = films.find((film) => film.id === filmId);
+
   return (
     <div className="w-100 d-flex justify-content-between align-items-start">
       <div className="w-40 d-flex justify-content-start align-items-start">
-        {films
-          .filter((film) => film.id === filmId)
-          .map((film) => (
-            <Card className="w-100 ">
-              <Card.Img variant="top" className="w-100" src={film.src} />
-              <Card.Body>
-                <Card.Title className="fs-3 ">{film.title}</Card.Title>
-                <Card.Text className="fs-6 truncate-text">
-                  {film.desc}
-                </Card.Text>
-              </Card.Body>
-            </Card>
-          ))}
+        {film ? (
+          <Card className="w-100 ">
+            <Card.Img variant="top" className="w-100" src={film.src} />
+            <Card.Body>
+              <Card.Title className="fs-3 ">{film.title}</Card.Title>
+              <Card.Text className="fs-6 truncate-text">{film.desc}</Card.Text>
+            </Card.Body>
+          </Card>
+        ) : null}
       </div>
       <div className="w-50 d-flex flex-column justify-content-center align-items-center">
         <Hall />
