--- src/pages/LoginP.tsx
@@ -1,8 +1,0 @@
-import { FC } from "react";
-import Login from "../components/Login/Login";
-
-const LoginP: FC = () => {
-  return <Login />;
-};
-
-export default LoginP;
