+++ src/routes.tsx
@@ -1,7 +1,7 @@
 import AboutFilm from "./pages/AboutFilm";
 import AboutUS from "./pages/AboutUS";
 import Admin from "./pages/Admin";
-import LoginP from "./pages/LoginP";
+import LoginP from "./pages/Login";
 import Basket from "./pages/BasketPage";
 import Films from "./pages/Films";
 import BuyTicket from "./pages/BuyTicket";
