+++ src/utils/constsRoles.ts
@@ -1,0 +1,2 @@
+export const ADMIN = "ADMIN";
+export const USER = "USER";
